fix 스터디 활성화 종료시 수정 못 하도록 변경 (프론트에서 함수 지우고 백에서도 작동 막음)

+++ reviews/views.py
@@ -41,52 +41,72 @@ def create(request):
     return render(request, 'reviews/form.html', context)
 
 def detail(request, study_pk):
-    study = Study.objects.get(pk=study_pk)
     review_form = ReviewForm()
+    study = Study.objects.get(pk=study_pk)
     cnt = len(Accepted.objects.filter(study=study))
+    users = Accepted.objects.filter(study_id=study_pk)
+    for user in users:
+        if user.users == request.user:
+            user_accepted = True
+            break
+    else:
+        user_accepted = False
     context = {'study' : study,
                'review_form': review_form,
-               'cnt':cnt}
+               'cnt':cnt,
+               'check':user_accepted
+               }
     return render(request,'reviews/detail.html', context)
 
 def userlist(request, study_pk):
     users = Accepted.objects.filter(study_id=study_pk)
-    study = Study.objects.filter(pk=study_pk)
+    study = Study.objects.get(pk=study_pk)
+    cnt = len(Accepted.objects.filter(study=study))
+    for user in users:
+        if user.users == request.user:
+            user_accepted = True
+            break
+    else:
+        user_accepted = False
     context = {
         'members':users,
-        'study':Study.objects.get(pk=study_pk),
+        'study':study,
+        'cnt':cnt,
+        'check':user_accepted
     }
     return render(request, 'reviews/userlist.html', context)
 
 
 def update(request, study_pk):
     study = Study.objects.get(pk=study_pk)
-    if request.user == study.host:
-        if request.method == 'POST':
-            study_form = StudyForm(request.POST, request.FILES, instance=study)
-            if study_form.is_valid():
-                study = study_form.save(commit=False)
-                study.categorie = request.POST['categorie']
-                study.study_type = request.POST['study_type']
-                study.location_type = request.POST['location_type']
-                study.location = request.POST['location']
-                study.X = request.POST['X']
-                study.Y = request.POST['Y']
-                study.deadline = request.POST['deadline']
-                study.host = request.user
-                study.save()
-                study_form.save()
-                return redirect('reviews:detail', study_pk)
+    if study.isactive:
+        if request.user == study.host:
+            if request.method == 'POST':
+                study_form = StudyForm(request.POST, request.FILES, instance=study)
+                if study_form.is_valid():
+                    study = study_form.save(commit=False)
+                    study.categorie = request.POST['categorie']
+                    study.study_type = request.POST['study_type']
+                    study.location_type = request.POST['location_type']
+                    study.location = request.POST['location']
+                    study.X = request.POST['X']
+                    study.Y = request.POST['Y']
+                    study.deadline = request.POST['deadline']
+                    study.host = request.user
+                    study.save()
+                    study_form.save()
+                    return redirect('reviews:detail', study_pk)
+            else:
+                study_form = StudyForm(instance=study)
+            context = {'study_form': study_form}
+            return render(request, 'reviews/form.html', context)
         else:
-            study_form = StudyForm(instance=study)
-        context = {'study_form': study_form}
-        return render(request, 'reviews/form.html', context)
-    else:
-        return redirect('reviews:detail', study_pk)
+            return redirect('reviews:detail', study_pk)
 
 def delete(request, study_pk):
     study = Study.objects.get(pk=study_pk)
-    study.delete()
+    if study.isactive:
+        study.delete()
     return redirect('reviews:index')
 
 
@@ -95,43 +115,50 @@ def join(request, study_pk, user_pk):
     study = Study.objects.get(pk=study_pk)
     accepted = Accepted.objects.filter(study_id=study_pk)
     users = Accepted.objects.filter(users_id=user_pk)
-    print(users)
-    if study.limits > len(accepted):
-        for joined in users:
-            if joined in accepted:
-                print('이미 가입되어 있습니다.')
+    if study.isactive:
+        if study.limits > len(accepted):
+            for joined in users:
+                if joined in accepted:
+                    print('이미 가입되어 있습니다.')
+                    return redirect('reviews:detail', study_pk)
+            else:
+                Aform = Accepted(joined=False,study=study,users=request.user)
+                Aform.save()
+                print('가입 신청')
                 return redirect('reviews:detail', study_pk)
         else:
-            Aform = Accepted(joined=False,study=study,users=request.user)
-            Aform.save()
-            print('가입 신청')
             return redirect('reviews:detail', study_pk)
     else:
-        return redirect('reviews:detail', study_pk)
+        return redirect('reviews:index')
 
 def study_accepted(request, study_id, users_id):
     study = Study.objects.get(id=study_id)
     user = User.objects.get(id=users_id)
     aform = Accepted.objects.get(users=user, study=study)
-    if request.user == study.host:
-        aform.joined = True
-        aform.save()
-        return redirect('reviews:userlist', study_id)
+    if study.isactive:
+        if request.user == study.host:
+            aform.joined = True
+            aform.save()
+            return redirect('reviews:userlist', study_id)
+        else:
+            return redirect('reviews:userlist', study_id)
     else:
-        return redirect('reviews:userlist', study_id)
+        return redirect('reviews:index')
 
 def study_kick(request, study_id, users_id):
     study = Study.objects.get(id=study_id)
     user = User.objects.get(id=users_id)
     aform = Accepted.objects.get(users=user, study=study)
-    if request.user == study.host and user != study.host:
-        aform.delete()
-        return redirect('reviews:userlist', study_id)
-    elif request.user == user and user != study.host:
-        aform.delete()
-        return redirect('reviews:index')
-    else:
-        return redirect('reviews:userlist', study_id)
+    if study.isactive:
+        if request.user == study.host and user != study.host:
+            aform.delete()
+            return redirect('reviews:userlist', study_id)
+        elif request.user == user and user != study.host:
+            aform.delete()
+            return redirect('reviews:userlist', study_id)
+        else:
+            return redirect('reviews:userlist', study_id)
+    return redirect('reviews:index')
 
 url="https://kapi.kakao.com/v2/api/talk/memo/default/send"
 
@@ -154,4 +181,10 @@ def gathering(request, study_pk):
             print(str(response.json()))
         else:
             print('no')
+    return redirect('reviews:detail', study_pk)
+
+def done(request, study_pk):
+    study = Study.objects.get(pk=study_pk)
+    study.isactive = False
+    study.save()
     return redirect('reviews:detail', study_pk)
